requirements.txt
gradio
fastAPI
uvicorn
jinja2
itsdangerous
greenlet
gevent
pycryptodome
pyexecjs